% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component-disease_indicator.R
\name{disease-indicator}
\alias{disease-indicator}
\alias{diseaseIndicator}
\alias{updateDiseaseIndicator}
\title{Disease indicator progress bar}
\usage{
diseaseIndicator(id, label, value)

updateDiseaseIndicator(
  session = getDefaultReactiveDomain(),
  id,
  value,
  label = NULL,
  palette = NULL
)
}
\arguments{
\item{id}{An id used to update the progress bar.}

\item{label}{Label to be displayed within progress bar}

\item{value}{Value of the progress bar between 0 and 100}

\item{session}{The 'session' object passed to function given to shinyServer.}
}
\value{
A progress bar that can be added to a UI definition.
}
\description{
Create a progress bar to visualise disease indicator
}
\examples{
if (interactive()) {
  library("shiny")

  ui <- fluidPage(
    fluidRow(
      column(
        width = 3,
        diseaseIndicator("lethality", "lethality", value = 55)
      )
    )
  )

  server <- function(input, output, session) {
    new_value <- reactive({
      invalidateLater(3000)
      sample(1:100, 1)
    })

    observeEvent(new_value(), {
      updateDiseaseIndicator(session, id = "lethality", value = new_value())
    })
  }
  shinyApp(ui = ui, server = server)
}
}
