% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-formatNumber.R
\name{formatNumbers}
\alias{formatNumbers}
\title{Format Numbers}
\usage{
formatNumbers(x, accuracy = 1)
}
\arguments{
\item{x}{number to be formatted}

\item{accuracy}{accuracy parameter}
}
\description{
Format Numbers
}
