% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic-spread_functions.R
\name{getAirborneSpreadChance}
\alias{getAirborneSpreadChance}
\title{Calculates airborne spread chance}
\usage{
getAirborneSpreadChance(infectiousness, airborne_bonus = 0)
}
\arguments{
\item{infectiousness}{number}

\item{infected_proportion}{proportion of infected population}
}
\description{
Calculates airborne spread chance
}
