% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic-spread_functions.R
\name{getCrossCountrySpreadChance}
\alias{getCrossCountrySpreadChance}
\title{Calculates cross-country spread chance}
\usage{
getCrossCountrySpreadChance(infectiousness, infected_proportion)
}
\arguments{
\item{infectiousness}{number}

\item{infected_proportion}{proportion of infected population}
}
\description{
Calculates cross-country spread chance
}
