% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic-death_chance.R
\name{getDeathChance}
\alias{getDeathChance}
\title{Calculates death chance}
\usage{
getDeathChance(lethality)
}
\arguments{
\item{infectiousness}{number}

\item{infected_proportion}{proportion of infected population}
}
\description{
Calculates death chance
}
