% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic-spread_functions.R
\name{getInCountrySpreadChance}
\alias{getInCountrySpreadChance}
\title{Calculates in-country spread chance}
\usage{
getInCountrySpreadChance(infectiousness, infected_proportion)
}
\arguments{
\item{infectiousness}{number}

\item{infected_proportion}{proportion of infected population}
}
\description{
Calculates in-country spread chance
}
