% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic-recovery_functions.R
\name{getRecoveryChance}
\alias{getRecoveryChance}
\title{Calculates in-country spread chance}
\usage{
getRecoveryChance(medical_progress)
}
\arguments{
\item{medical_progress}{gameState parameter}
}
\description{
Calculates in-country spread chance
}
